+++ shotmanager/utils/utils_ui.py
@@ -175,7 +175,6 @@ class UAS_ShotManager_OpenFileBrowser(Operator, ImportHelper):
 # Dev note: This function has to be here for the moment cause it is passed
 # in stampinfo code to a call to uas_stamp_info.querybox
 def reset_render_properties():
-    print("reset_render_properties")
     props = bpy.context.scene.UAS_shot_manager_props
     props.reset_render_properties()